requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
requests>=2.31.0

# FastAPI 관련 패키지 (PC 전용)
fastapi>=0.104.0       # API 서버 프레임워크
uvicorn[standard]>=0.24.0  # ASGI 서버 (FastAPI 실행)
pydantic>=2.0.0        # 데이터 검증 및 직렬화
python-dotenv>=1.0.0   # 환경 변수 관리

